refactor(dashboard): separate gql queries and mutations
