+++ dashboard/src/pages/createPost.page.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import { useMutation } from '@apollo/client';
 import { Redirect } from 'react-router-dom';
 import Loader from '../components/loader';
-import { CREATE_POST, CreatePostResult } from '../graphql/queries';
 import PostForm from '../components/postForm';
+import { CREATE_POST, CreatePostResult } from '../graphql/post.mutations';
 
 export default function CreatePostPage() {
   const [createPost, { data, loading }] = useMutation<CreatePostResult, { title: string; body: string }>(CREATE_POST);
